Refactor/structure (#37)
* refactor: component

* refactor: firebase

* refactor: utils

* refactor: layout

* fix: eslint

* fix: eslint

* fix: packges

* fix: eslint

* fix: file name

* fix: package.json
