+++ package.json
@@ -4,11 +4,11 @@
   "description": "",
   "main": "index.js",
   "scripts": {
-    "dev": "env-cmd -f config/.env.local next",
+    "dev": "next",
     "build": "next build",
     "start": "next start",
-    "lint": "eslint .",
-    "lint:fix": "eslint . --fix",
+    "lint": "next lint",
+    "lint:fix": "next lint --fix",
     "format": "prettier --write . --config config/.prettierrc.json --ignore-path config/.prettierignore",
     "prepare": "husky install",
     "pre-commit": "lint-staged -c config/.lintstagedrc.json && npm run lint && npm run type-check",
@@ -33,7 +33,7 @@
     "@mui/material": "^5.6.4",
     "firebase": "^9.7.0",
     "firebase-admin": "^11.0.1",
-    "next": "^12.2.4",
+    "next": "^12.2.5",
     "nookies": "^2.5.2",
     "react": "^18.1.0",
     "react-dom": "^18.1.0",
@@ -49,9 +49,8 @@
     "@typescript-eslint/eslint-plugin": "^5.23.0",
     "@typescript-eslint/parser": "^5.23.0",
     "autoprefixer": "^10.4.6",
-    "env-cmd": "^10.1.0",
     "eslint": "8.14.0",
-    "eslint-config-next": "^12.2.4",
+    "eslint-config-next": "^12.2.5",
     "eslint-config-prettier": "^8.5.0",
     "eslint-plugin-import": "^2.26.0",
     "eslint-plugin-react": "^7.29.4",
