+++ src/components/Icon/index.tsx
