--- src/components/splash/index.tsx
@@ -1,5 +1,0 @@
-const Splash = () => {
-  return <>loading</>;
-};
-
-export default Splash;
