+++ src/lib/firebase/client/app.ts
