+++ src/lib/firebase/client/auth/index.ts
