+++ src/lib/firebase/client/firestore/functions.ts
