+++ src/lib/firebase/client/index.ts
