+++ src/pages/signin/index.tsx
@@ -2,10 +2,10 @@ import Image from "next/image";
 import Link from "next/link";
 import { useRouter } from "next/router";
 import { useCallback } from "react";
-import bgMahjong1Image from "@images/bg/mahjong1.jpeg";
-import GoogleIconNormalSvg from "@svgs/google/btn_google_light_normal_ios.svg";
-import { NextPageWithLayout } from "@types";
+import bgMahjong1Image from "~/assets/images/bg/mahjong1.jpeg";
+import GoogleIconNormalSvg from "~/assets/svgs/google/btn_google_light_normal_ios.svg";
 import { METHOD } from "~/services/auth";
+import { NextPageWithLayout } from "~/types";
 
 const Signin: NextPageWithLayout = () => {
   const router = useRouter();
