+++ src/pages/signin/redirect.tsx
@@ -1,10 +1,10 @@
 import { getRedirectResult } from "firebase/auth";
 import router from "next/router";
 import { useEffect } from "react";
-import { authTokenCookie } from "@lib/cookie";
-import { NextPageWithLayout } from "@types";
-import { auth } from "~/firebase/client";
+import { authTokenCookie } from "~/lib/cookie";
+import { auth } from "~/lib/firebase/client";
 import { Method, METHOD, signin } from "~/services/auth";
+import { NextPageWithLayout } from "~/types";
 
 const Signin: NextPageWithLayout = () => {
   const onLoad = async () => {
