+++ src/types/index.ts
@@ -1,13 +1,6 @@
 import { NextPage } from "next";
 import { ReactElement, ReactNode } from "react";
 
-export type Auth = {
-  loadingUser: boolean;
-  signedIn: boolean;
-  signOut: VoidFunction;
-  user: User | undefined;
-};
-
 export type User = {
   uid: string;
   name: string | null;
